requirement.txt
Django==1.11.8
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2018.7
